(SUP-3101)
This commit modifies S0022
Now fact pe_status_check.S0022 will only exist if there is a license
file on the machine. Previously the factwould resolve to a default
value of true.

Now pe_status_check.S0022 will resolve to true if the license file
is valid and does not expire within the next 90 days or if it's a
Perpetual license.
pe_status_check.S0022 will resolve to false if the license will
expire within 90 days, if the license file contains an invalid
end date, or if the license file does not contain a valid
Perpetual or Subscription license_type.

This commit also modifies the spec test for S0022. Previously the spec
test expected failure if the license file doesn't exist on the machine.
Now the spec test adds a new temporary license file with an end_date
that has already passed for the failure condition.