(SUP-4204) Add lastrunreport check
Added in agent_status_check indicator AS005 to review the lastrunreport
for any logs at Warning or Error level. The indicator resolves to false
if there is at least one Warning or Error. This indicator is helpful to
capture errors that might not cause the Puppet run to fail such as
Facter errors.